feat - Hero image, more content
I added the hero image. feel free to replace with a better one if it is not to your liking. I also added more filler content that you can go in and edit to be relavent information to you for your portfolio.
I added the css for the text on the hero image.